pyproject.toml
[project]
name = "strava-viewer"
version = "1.0.0"
description = "Python Package to run Strava Viewer website"
authors = [{ name = "Malo Virmoux", email = "malo.virmoux.2@gmail.com" }]
license = { text = "ISC" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi",
    "dotenv",
    "flask",
    "flask_cors",
    "flask-login",
    "polars",
    "psycopg2",
    "requests",
    "polars",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
